+++ src/searches/common/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(searches_common STATIC
     children.cpp
     heuristics.cpp
+    tools.cpp
 )
 target_include_directories(searches_common PUBLIC ${CMAKE_SOURCE_DIR})
