+++ src/searches/common/searches_common.h
@@ -2,4 +2,4 @@
 #include "children.h"
 #include "heuristics.h"
 #include "retval.h"
-
+#include "tools.h"
