+++ src/searches/common/tools.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <utility>
+
+
+
+// x,y to a single int
+int xy2int(int x, int y, int map_size);
+std::pair<int, int> int2xy(int index, int map_size);
