+++ src/searches/fringe_search.cpp
@@ -2,15 +2,6 @@
 #include <iomanip>
 #include <iostream>
 
-int xy2int(int x, int y, int map_size) {
-  return y*map_size + x;
-}
-
-std::pair<int, int> int2xy(int index, int map_size) {
-  int x = index % map_size;
-  int y = index / map_size;
-  return std::make_pair(x,y);
-}
 
 RetVal fringe_search(int startx, int starty, int goalx, int goaly, std::vector<std::string> citymap) {
